feat(fe): remove create problem alert dialog
